requirements.txt
pip~=22.1.2
wheel~=0.36.2
setuptools~=49.2.1
python-dotenv~=0.20.0
psycopg2-binary~=2.8.6
Django==2.2.19
Pillow==9.2.0
djangorestframework~=3.13.1